Menu fixes & problems
I fixed the lag and unresponsiveness of the menus -- turned out I wasn't supposed to call so many things in HUD blueprints, but rather in the individual widget blueprints.

New bugs have come up:
1) The stats aren't setting after character creation (i.e. speed is messed up)
2) The title screen menu won't go away after loading a new level, no idea why

However this has made me think about things more and I want to rework the stat system so that all your stats start at 1 and you can choose a stat to raise on levelups, which is more intuitive. I'll implement that next.